🚨 #74 fix: 서버 컨텍스트에서 createServerClient 를 사용할 수 있음
- Error:   × You're importing a component that needs "next/headers". That only works in a Server Component which is not supported in the pages/ directory. Read more: https://nextjs.org/docs/app/building-your-application/rendering/server-components

Import trace for requested module:
./app/lib/queries/getBooks.ts
